Move data classes used only in the local new session queue to that package
